Move protos out of the protos directory.
