--- ludwig/profiling/__init__.py
@@ -1,1 +1,0 @@
-from ludwig.profiling.python import dataset_profile_pb2  # noqa
