+++ ludwig/profiling/gen-proto.sh
@@ -1,8 +1,8 @@
 protoc --go_out=. --python_out=. \
-    ludwig/profiling/proto/whylogs_messages.proto \
-    ludwig/profiling/proto/dataset_profile.proto \
+    ludwig/profiling/whylogs_messages.proto \
+    ludwig/profiling/dataset_profile.proto \
     --go_opt=paths=source_relative && {
-    mv ludwig/profiling/proto/*.pb.go ludwig/profiling/go/
+    mv ludwig/profiling/*.pb.go ludwig/profiling/go/
 } && {
-    mv ludwig/profiling/proto/*pb2.py ludwig/profiling/python
+    mv ludwig/profiling/*pb2.py ludwig/profiling/python
 }
